# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"HRMS Lite API","description":"A lightweight Human Resource Management System API","version":"1.0.0"},"paths":{"/api/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Check if the API is running","operationId":"health_check_api_health_get","responses":{"200":{"description":"API is healthy","content":{"application/json":{"schema":{}}}}}}},"/api/employees/":{"post":{"tags":["employees"],"summary":"Create Employee","description":"Create a new employee with optional photo upload","operationId":"create_employee_api_employees__post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_employee_api_employees__post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"400":{"description":"Invalid input or duplicate employee","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["employees"],"summary":"List Employees","description":"List all employees","operationId":"list_employees_api_employees__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Employee"},"title":"Response List Employees Api Employees  Get"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/employees/json":{"post":{"tags":["employees"],"summary":"Create Employee Json","description":"Create a new employee via JSON (no file upload).","operationId":"create_employee_json_api_employees_json_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"400":{"description":"Invalid input or duplicate employee","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/employees/{employee_id}":{"get":{"tags":["employees"],"summary":"Get Employee","description":"Get a specific employee by employee_id","operationId":"get_employee_api_employees__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Employee"}}}},"404":{"description":"Employee not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["employees"],"summary":"Delete Employee","description":"Delete an employee","operationId":"delete_employee_api_employees__employee_id__delete","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Employee not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/attendance/":{"post":{"tags":["attendance"],"summary":"Mark Attendance","description":"Mark or update attendance for an employee. Returns 201 for new, 200 for update.","operationId":"mark_attendance_api_attendance__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttendanceCreate"}}}},"responses":{"200":{"description":"Attendance updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Attendance"}}}},"201":{"description":"Attendance created"},"400":{"description":"Invalid input","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"404":{"description":"Employee not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["attendance"],"summary":"List Attendance","description":"List all attendance records","operationId":"list_attendance_api_attendance__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Attendance"},"title":"Response List Attendance Api Attendance  Get"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/attendance/employee/{employee_id}":{"get":{"tags":["attendance"],"summary":"Get Employee Attendance","description":"Get attendance records for a specific employee","operationId":"get_employee_attendance_api_attendance_employee__employee_id__get","parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"string","title":"Employee Id"}},{"name":"start_date","in":"query","required":false,"schema":{"type":"string","format":"date","title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"type":"string","format":"date","title":"End Date"}},{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Attendance"},"title":"Response Get Employee Attendance Api Attendance Employee  Employee Id  Get"}}}},"404":{"description":"Employee not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"500":{"description":"Server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["root"],"summary":"Read Root","description":"API root endpoint","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Attendance":{"properties":{"employee_id":{"type":"string","minLength":1,"title":"Employee Id"},"date":{"type":"string","format":"date","title":"Date"},"status":{"$ref":"#/components/schemas/AttendanceStatusEnum"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["employee_id","date","status","id","created_at","updated_at"],"title":"Attendance"},"AttendanceCreate":{"properties":{"employee_id":{"type":"string","minLength":1,"title":"Employee Id"},"date":{"type":"string","format":"date","title":"Date"},"status":{"$ref":"#/components/schemas/AttendanceStatusEnum"}},"type":"object","required":["employee_id","date","status"],"title":"AttendanceCreate"},"AttendanceStatusEnum":{"type":"string","enum":["Present","Absent"],"title":"AttendanceStatusEnum"},"Body_create_employee_api_employees__post":{"properties":{"employee_id":{"type":"string","title":"Employee Id"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"department":{"type":"string","title":"Department"},"photo":{"type":"string","format":"binary","title":"Photo"}},"type":"object","required":["employee_id","full_name","email","department"],"title":"Body_create_employee_api_employees__post"},"Employee":{"properties":{"employee_id":{"type":"string","maxLength":50,"minLength":1,"title":"Employee Id"},"full_name":{"type":"string","maxLength":100,"minLength":1,"title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"department":{"type":"string","maxLength":100,"minLength":1,"title":"Department"},"photo_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Path"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["employee_id","full_name","email","department","id","created_at","updated_at"],"title":"Employee"},"EmployeeCreate":{"properties":{"employee_id":{"type":"string","maxLength":50,"minLength":1,"title":"Employee Id"},"full_name":{"type":"string","maxLength":100,"minLength":1,"title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"department":{"type":"string","maxLength":100,"minLength":1,"title":"Department"},"photo_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Path"}},"type":"object","required":["employee_id","full_name","email","department"],"title":"EmployeeCreate"},"ErrorDetail":{"properties":{"detail":{"type":"string","title":"Detail"}},"type":"object","required":["detail"],"title":"ErrorDetail"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}